+++ pkg/analyzer/demographics.go
@@ -13,18 +13,18 @@ type Demographics struct {
 
 // Languages returns the visitor count grouped by language.
 func (demographics *Demographics) Languages(filter *Filter) ([]model.LanguageStats, error) {
-	q, args := demographics.analyzer.selectByAttribute(filter, FieldLanguage)
-	return demographics.store.SelectLanguageStats(q, args...)
+	ctx, q, args := demographics.analyzer.selectByAttribute(filter, FieldLanguage)
+	return demographics.store.SelectLanguageStats(ctx, q, args...)
 }
 
 // Countries returns the visitor count grouped by country.
 func (demographics *Demographics) Countries(filter *Filter) ([]model.CountryStats, error) {
-	q, args := demographics.analyzer.selectByAttribute(filter, FieldCountry)
-	return demographics.store.SelectCountryStats(q, args...)
+	ctx, q, args := demographics.analyzer.selectByAttribute(filter, FieldCountry)
+	return demographics.store.SelectCountryStats(ctx, q, args...)
 }
 
 // Cities returns the visitor count grouped by city.
 func (demographics *Demographics) Cities(filter *Filter) ([]model.CityStats, error) {
-	q, args := demographics.analyzer.selectByAttribute(filter, FieldCity, FieldCountryCity)
-	return demographics.store.SelectCityStats(q, args...)
+	ctx, q, args := demographics.analyzer.selectByAttribute(filter, FieldCity, FieldCountryCity)
+	return demographics.store.SelectCityStats(ctx, q, args...)
 }